docs/legal-chatbot-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
python-multipart==0.0.12

